Allow install.sh to bootstrap itself
This allows OSXConfig install from just the one script file alone.

Lots of snarky documentation because this is a low-value script and the
changes were written three years ago, never run since then, and only now
attempted to understand. Of course, real code gets documented when
written, not made a mess like this. :)